feat: implemented 2 version for cross-chain modal
